Compartmentalizing more SVG classes
